+++ Core/Layer/Worlds/WorldLayer.cs
@@ -252,7 +252,7 @@ public partial class WorldLayer : IGameLayerParent
         if (worldModel == null)
             return;
 
-        config.ApplyConfiguration(worldModel.ConfigValues, writeToConfig: false);
+        config.ApplyConfiguration(worldModel.ConfigValues);
     }
 
     public static SinglePlayerWorld? CreateWorldGeometry(GlobalData globalData, IConfig config, IAudioSystem audioSystem,