+++ Tests/Unit/GameAction/DoorClip.cs
@@ -4,6 +4,7 @@ using Helion.World.Entities.Players;
 using Helion.World.Impl.SinglePlayer;
 using Helion.World.Physics;
 using Xunit;
+using Helion.Util.Configs.Components;
 
 namespace Helion.Tests.Unit.GameAction;
 
@@ -25,7 +26,7 @@ public class DoorClip
     [Fact(DisplayName = "Door ceiling clips floor")]
     public void DoorClipsFloor()
     {
-        World.Config.Compatibility.VanillaSectorPhysics.Value.Should().BeFalse();
+        World.Config.Compatibility.VanillaSectorPhysics.Value.ToBool().Should().BeFalse();
         var sector = GameActions.GetSector(World, 2);
         sector.Ceiling.Z.Should().Be(0);
         GameActions.ActivateLine(World, Player, 3, ActivationContext.UseLine).Should().BeTrue();